refactor: move execInfo, status to separate files
simplify arg passed to parallelExec
